requirements.txt
openai
langchain
langchain-community
langchain-openai
fastapi
uvicorn
gunicorn
youtube-transcript-api
tiktoken
wikipedia
python-dotenv
faiss-cpu
deep-translator
pytube
requests
